REFACTOR-#6815: move experimental parsers into 'modin.experimental' folder
