+++ modin/core/execution/ray/implementations/pandas_on_ray/io/io.py
@@ -30,9 +30,6 @@ from modin.core.io import (
     SQLDispatcher,
 )
 from modin.core.storage_formats.pandas.parsers import (
-    ExperimentalCustomTextParser,
-    ExperimentalPandasPickleParser,
-    PandasCSVGlobParser,
     PandasCSVParser,
     PandasExcelParser,
     PandasFeatherParser,
@@ -47,6 +44,11 @@ from modin.experimental.core.io import (
     ExperimentalCustomTextDispatcher,
     ExperimentalPickleDispatcher,
     ExperimentalSQLDispatcher,
+)
+from modin.experimental.core.storage_formats.pandas.parsers import (
+    ExperimentalCustomTextParser,
+    ExperimentalPandasPickleParser,
+    PandasCSVGlobParser,
 )
 
 from ..dataframe import PandasOnRayDataframe
