+++ modin/core/execution/unidist/implementations/pandas_on_unidist/io/io.py
@@ -29,9 +29,6 @@ from modin.core.io import (
     SQLDispatcher,
 )
 from modin.core.storage_formats.pandas.parsers import (
-    ExperimentalCustomTextParser,
-    ExperimentalPandasPickleParser,
-    PandasCSVGlobParser,
     PandasCSVParser,
     PandasExcelParser,
     PandasFeatherParser,
@@ -46,6 +43,11 @@ from modin.experimental.core.io import (
     ExperimentalCustomTextDispatcher,
     ExperimentalPickleDispatcher,
     ExperimentalSQLDispatcher,
+)
+from modin.experimental.core.storage_formats.pandas.parsers import (
+    ExperimentalCustomTextParser,
+    ExperimentalPandasPickleParser,
+    PandasCSVGlobParser,
 )
 
 from ..dataframe import PandasOnUnidistDataframe
